% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{easy_exact_limits}
\alias{easy_exact_limits}
\alias{easy_exact_xlim}
\alias{easy_exact_ylim}
\title{Easily specify exact limits}
\usage{
easy_exact_limits(which = c("x", "y"), limits = NULL, expand = c(0, 0,
  0.1, 0), teach = FALSE)

easy_exact_xlim(..., teach = FALSE)

easy_exact_ylim(..., teach = FALSE)
}
\arguments{
\item{which}{"x" or "y"; which limits are to be adjusted?}

\item{limits}{a two-element numeric-like vector specifying the lower and upper bounds of the axis extent}

\item{expand}{a four-element numeric-like vector specifying the multiplicative and additive expansion parameters. See Details.}

\item{teach}{print longer equivalent ggplot expression?}
}
\value{
a \code{\link[ggplot2]{theme}} object  which can be used in
\code{\link[ggplot2]{ggplot2}} calls
}
\description{
Easily specify exact limits
}
\examples{

ggplot(mtcars) +
  geom_bar(aes(x = factor(cyl))) +
  easy_exact_limits("y")

ggplot(mtcars) +
  geom_bar(aes(x = factor(cyl))) +
  easy_exact_limits("y", limits = c(-10, NA))

ggplot(mtcars) +
  geom_bar(aes(x = factor(cyl))) +
  easy_exact_ylim()

## still have control; set the limits to the extent with no expansion
ggplot(mtcars) +
    geom_point(aes(x = mpg, y = hp)) +
    easy_exact_xlim(expand = c(0, 0, 0, 0), limits = c(min(mtcars$mpg), max(mtcars$mpg))) +
    easy_exact_ylim(expand = c(0, 0, 0, 0), limits = c(min(mtcars$hp), max(mtcars$hp)))

## teach how to do this
ggplot(mtcars) +
  geom_bar(aes(x = factor(cyl))) +
  easy_exact_ylim(teach = TRUE)
}
\author{
Jonathan Carroll
}
