% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticks-text.r
\name{easy_flush_ticks}
\alias{easy_flush_ticks}
\title{Makes axis text labels flush on the ends}
\usage{
easy_flush_ticks(gg, flush = "XY", teach = TRUE, plot = TRUE)
}
\arguments{
\item{gg}{ggplot2 plot object}

\item{flush}{either "\code{X}" or "\code{Y}" or "\code{XY}" to flush individual or both axes. Default: both.}

\item{teach}{if \code{TRUE} then display \code{theme()} statements and copy them to the clipboard}

\item{plot}{if \code{FALSE} then the ggplot object will be returned \emph{invisibly}}
}
\value{
ggplot2 object with \code{theme()} elements added
}
\description{
A covenience function intended for basic, fixed-scale plots only (i.e. does not handle
free scales in facets).\cr
\cr
You need to pass in a ggplot2 object to this function. It can't be \code{+}'d in a chain
of geoms, coords, scales, themes, etc. It also builds the plot (but does not display
it) so if the plt takes a while (i.e. has lots of data or transforms) this will also
take a while.
}
\note{
Intended for basic, fixed-scale plots only (i.e. does not handle free scales in facets).
Also, the clipboard operations will only occur if \code{clipr} is installed.
}
\examples{
library(ggplot2)

# Flush ticks' text
p <- ggplot(mtcars, aes(wt, mpg, colour = cyl, size = hp)) +
  geom_point()
easy_flush_ticks(p)
}
