% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beautify.R
\name{easy_grid}
\alias{easy_grid}
\title{So many grid options... So little time}
\usage{
easy_grid(color, linetype, axis = "both", major.only = FALSE)
}
\arguments{
\item{color}{the color of the grid}

\item{linetype}{the type of line for the grid}

\item{axis}{"x" for x-axis only, "y" for y-axis only, "both" for both}

\item{major.only}{should only the major tick marks correspond with grid lines}
}
\value{
the desired grid
}
\description{
So many grid options... So little time
}
\examples{
plt + easy_grid("white", "dashed")
}
