% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_labs.R
\name{easy_labs}
\alias{easy_labs}
\title{Create ggplot labels using label attribtute of data.frame column}
\usage{
easy_labs(...)
}
\arguments{
\item{...}{Arguments passed on to \code{ggplot2::labs}
\describe{
  \item{}{}
}}
}
\description{
Applies same logic as \code{\link[ggplot2]{labs}} but uses as default
the column label attribute if present as the variable label in the plot.
}
\examples{

lbl <- c('Sepal Length','Sepal Width','Petal Length','Petal Width','Flower Species')

iris_labs <- iris\%>\%label(vars=names(iris),labels = lbl)

p <- iris_labs\%>\%
  ggplot(aes(x=Sepal.Length,y=Sepal.Width))+
  geom_line(aes(colour=Species))
  
 p + easy_labs()
 
 p + easy_labs(title='my title',subtitle='mysubtitle',x='x axis label')


}
