% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_bar.R
\name{easy_range}
\alias{easy_range}
\title{Easy range bar plot}
\usage{
easy_range(df, fct_var, num_var, col = "brown1", teach = FALSE)
}
\arguments{
\item{df}{the data frame}

\item{fct_var}{the factor variable}

\item{num_var}{the function will calculate the range of num_var for each value of fct_var}

\item{col}{the color of the range bar plot borders}

\item{teach}{print longer equivalent \code{\link[ggplot2]{ggplot2}}
expression?}
}
\value{
a \code{\link[ggplot2]{ggplot2}} object
}
\description{
Create a range plot in the for of error bars that displays the range of a numeric variable according to the values of a factor variable
}
\examples{
library(ggplot2)
easy_range(df = mpg, fct_var = class, num_var = hwy)



}
\author{
Mohamed El Fodil Ihaddaden
}
