% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{easy_remove_legend_title}
\alias{easy_remove_legend_title}
\title{Easily remove legend title
Remove the legend title}
\usage{
easy_remove_legend_title(teach = FALSE)
}
\arguments{
\item{teach}{print longer equivalent \code{\link[ggplot2]{ggplot2}}
expression?}
}
\value{
a \code{\link[ggplot2]{theme}} object
}
\description{
Easily remove legend title
Remove the legend title
}
\examples{
library(ggplot2)
# remove legend title from all aesthetics
ggplot(mtcars, aes(wt, mpg, colour = cyl)) +
  geom_point() + easy_remove_legend_title()
}
