% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_labels.R
\name{easy_rotate_facet_labels}
\alias{easy_rotate_facet_labels}
\alias{easy_rotate_axis_labels}
\title{Rotate labels for axes and facets}
\usage{
easy_rotate_facet_labels(angle = 45, direction = "x",
  remove_label_background = TRUE)

easy_rotate_axis_labels(angle = 90, direction = "x")
}
\arguments{
\item{angle}{The angle at which labels should be rotated}

\item{direction}{The axes along which the operations should be performed}

\item{remove_label_background}{Whether or not label backgrounds should be removed along
rotated label axes}
}
\value{
An object that can be added to a \link[ggplot2]{ggplot}
}
\description{
Rotate labels for axes and facets
}
