% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beautify.R
\name{fix_margins}
\alias{fix_margins}
\title{Easy way to adjust plot margins}
\usage{
fix_margins(margin_values = c(2, 2, 1, 1), units = "cm")
}
\arguments{
\item{margin_values}{the values to adjust the margin by, the default seems
to work nicely for me...}

\item{units}{the units of the margin adjustment}
}
\value{
a theme with the desired margin adjustments
}
\description{
Easy way to adjust plot margins
}
\examples{
plt + fix_margins()
}
