% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbot.R
\name{ggbot}
\alias{ggbot}
\title{Helper bot to change plot aesthetics}
\usage{
ggbot(...)
}
\arguments{
\item{...}{character vector containing lexical tokens (a command or a vector or words)}
}
\value{
a \code{\link[ggplot2]{theme}} object which can be used in
\code{\link[ggplot2]{ggplot2}} calls.
}
\description{
Helper bot to change plot aesthetics
}
\details{
Commands currently supported:
\itemize{
\item change colour of points
\item remove axis, legend, everything
}
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(cyl, hp, col = "red")) +
  geom_point() +
  ggbot("blue points")
}
