% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggeasy-package.r
\docType{package}
\name{ggeasy}
\alias{ggeasy-package}
\alias{ggeasy}
\title{ggeasy.}
\description{
Helper functions for making using ggplot2 easier.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonocarroll/ggeasy}
  \item \url{https://jonocarroll.github.io/ggeasy/}
  \item Report bugs at \url{https://github.com/jonocarroll/ggeasy/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Carroll \email{rpkg@jcarroll.com.au} (\href{https://orcid.org/0000-0002-1404-5264}{ORCID})

Authors:
\itemize{
  \item Alicia Schep \email{aschep@gmail.com} (\href{https://orcid.org/0000-0002-3915-0618}{ORCID})
  \item Jonathan Sidi \email{yonicd@gmail.com} (\href{https://orcid.org/0000-0002-4222-1819}{ORCID})
}

Other contributors:
\itemize{
  \item Bob Rudis \email{bob@rud.is} (\href{https://orcid.org/0000-0001-5670-2640}{ORCID}) [contributor]
  \item Mohamed El Fodil Ihaddaden \email{ihaddaden.fodeil@gmail.com} [contributor]
  \item Thomas Neitmann \email{th.neitmann@gmail.com} [contributor]
}

}
\keyword{internal}
