% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{Add label attribute to data.frame columns}
\usage{
label(.data, vars, labels)
}
\arguments{
\item{.data}{data.frame, data.frame to manipulate}

\item{vars}{character, names of columns to add label attribute}

\item{labels}{character, labels to append to vars}
}
\value{
data.frame
}
\description{
Utility function to add label attribute to data.frame columns.
}
\examples{
lbl <- c('Sepal Length','Sepal Width','Petal Length','Petal Width','Flower Species')

iris_labs <- iris\%>\%label(vars=names(iris),labels = lbl)

}
\author{
Jonathan Sidi
}
