% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removal.R
\name{no_axis_labels}
\alias{no_axis_labels}
\title{Removes Axis Labels from GGPlot}
\usage{
no_axis_labels()
}
\value{
a ggplot theme with no axis labels
}
\description{
Removes Axis Labels from GGPlot
}
\examples{
plt + no_axis_labels()
}
