% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removal.R
\name{no_ticks}
\alias{no_ticks}
\title{Removes Axis Ticks and Tick Text from GGPlot}
\usage{
no_ticks()
}
\value{
a ggplot theme with no tick markings
}
\description{
Removes Axis Ticks and Tick Text from GGPlot
}
\examples{
plt + no_ticks()
}
