% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{panel_background_color}
\alias{panel_background_color}
\title{Change the Panel Area Color for GGPlot}
\usage{
panel_background_color(fill_color, outline_color)
}
\arguments{
\item{fill_color}{the fill -- internal area -- color}

\item{outline_color}{the outline -- border area -- color}
}
\value{
a ggplot theme with color scheme desired by user
}
\description{
Change the Panel Area Color for GGPlot
}
\examples{
plt + panel_background_color("black", "black")
}
