% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{plot_background_color}
\alias{plot_background_color}
\title{Change the Plot Area Color for GGPlot}
\usage{
plot_background_color(fill_color, outline_color)
}
\arguments{
\item{fill_color}{the fill -- internal area -- color}

\item{outline_color}{the outline -- border area  -- color}
}
\value{
a ggplot theme with desire color scheme by user
}
\description{
Change the Plot Area Color for GGPlot
}
\examples{
plt + plot_background_color("black", "black")
}
