% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_true_dark}
\alias{theme_true_dark}
\title{Create an actual dark theme for a plot}
\usage{
theme_true_dark()
}
\value{
a true dark theme
}
\description{
Create an actual dark theme for a plot
}
\examples{
plt + theme_true_dark()
}
