% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_bar.R
\name{easy_bar_median}
\alias{easy_bar_median}
\title{Easy median bar plot}
\usage{
easy_bar_median(
  df,
  fct_var,
  num_var,
  col = "brown1",
  fill = "brown4",
  teach = FALSE
)
}
\arguments{
\item{df}{the data frame}

\item{fct_var}{the factor variable}

\item{num_var}{the function will calculate the median of num_var for each value of fct_var}

\item{col}{the color of the bar plot borders}

\item{fill}{the filling color of the bar plot}

\item{teach}{print longer equivalent \code{\link[ggplot2]{ggplot2}}
expression?}
}
\value{
a \code{\link[ggplot2]{ggplot2}} object
}
\description{
Create a bar plot that displays the median of a numeric variable according to the values of a factor variable
}
\examples{
library(ggplot2)
easy_bar_median(df = mpg, fct_var = class, num_var = hwy)



}
\author{
Mohamed El Fodil Ihaddaden
}
