% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labs.R
\name{easy_labs}
\alias{easy_labs}
\title{Easily add ggplot labels using label attribtute of data.frame column}
\usage{
easy_labs(...)
}
\arguments{
\item{...}{A list of new name-value pairs. The name should either be an aesthetic, 
or one of "title", "subtitle", or "caption"}
}
\description{
Applies same logic as \code{\link[ggplot2]{labs}} but uses as default
the column label attribute if present as the variable label in the plot.
}
\examples{

lbl <- c('Sepal Length','Sepal Width','Petal Length','Petal Width','Flower Species')

iris_labs <- label(iris,vars=names(iris),labels = lbl)

p <- 
  ggplot(iris_labs,aes(x=Sepal.Length,y=Sepal.Width))+
  geom_line(aes(colour=Species))
  
 p 
  
 p + easy_labs()
 
 p + labs(title='my title') + easy_labs(subtitle='mysubtitle',x='x axis label')


}
